pyproject.toml
[project]
name = "fractal-guide"
version = "0.1.0"
description = ""
authors = [
    {name = "Guillermo Barquero",email = "gb@zriveapp.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai (>=2.3.0,<3.0.0)",
    "streamlit (>=1.50.0,<2.0.0)",
    "geopy (>=2.4.1,<3.0.0)",
    "streamlit-geolocation (>=0.0.10,<0.0.11)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
fractal-guide = "fractal_guide.src.main:main"
